travel-agent/requirements.txt
anima-labs>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
